pyproject.toml
[project]
name = "find-kor-routes"
version = "0.1.0"
description = "택시+대중교통 혼합 경로 탐색 MVP"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
find-route = "find_kor_routes.cli:main"
